fix signup auth with userauth.dbservice
