+++ src/controllers/Auth.controllers.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { AuthModule } from "../modules/Auth.module";
-import { UserDbService } from "../database/users/User.dbservice";
+// import { UserDbService } from "../database/users/User.dbservice";
 import { UserAuthDbService } from "../database/users/UserAuth.dbservice";
 import { User } from "../database/users/User.dbmodel";
 import { UserAuth } from "../database/users/UserAuth.dbmodel";
@@ -10,7 +10,7 @@ import  * as jwt from "jsonwebtoken";
 const router = express.Router();
 
 const authModule: AuthModule = new AuthModule();
-const usersDbService: UserDbService = new UserDbService();
+// const usersDbService: UserDbService = new UserDbService();
 const userAuthDbService: UserAuthDbService = new UserAuthDbService();
 
 // TODO: change this to be hidden
@@ -41,11 +41,11 @@ router.post('/users/create', async (req, res, next) => {
     console.log(req.body["publicAddress"]);
     console.log(req.body["name"]);
     // Check if the user is already created
-    let userModel: User = await usersDbService.selectOneRowByPrimaryId(req.body.publicAddress);
+    let userModel: User = await userAuthDbService.selectOneRowByPrimaryId(req.body.publicAddress);
     if (userModel == null) {
         // If not, create the user and send back to frontend
         console.log("user doesn't exist, creating user");
-        userModel = await usersDbService.createUser(req.body.publicAddress, req.body.name);
+        userModel = await userAuthDbService.createUser(req.body.publicAddress, req.body.name);
     }
     console.log("user exists, no need to create user"); // ADD SOME INDICATION IN RESPONSE
     console.log(userModel);
