+++ src/database/users/UserAuth.dbservice.ts
@@ -23,6 +23,19 @@ export class UserAuthDbService extends DatabaseService<UserAuth> {
     return UserAuth.dbRowToDbModel(dbRow);
   }
 
+  async createUser(address: string, name: string): Promise<UserAuth> | null {
+    let nonce = uuid();
+    const values = [address, name, nonce];
+    let users: UserAuth[] = await super.runParameterizedQueryWithValuesArray(
+      this.createUserQuery,
+      values
+    );
+    if (users === []) {
+      return null;
+    }
+    return users[0];
+  }
+
   async getUserAuth(publicAddress: string): Promise<UserAuth> | null {
     const values = [publicAddress];
     let users: UserAuth[] = await super.runParameterizedQueryWithValuesArray(
@@ -35,7 +48,15 @@ export class UserAuthDbService extends DatabaseService<UserAuth> {
     return users[0];
   }
 
+  private createUserQuery = {
+    text:
+      "insert into users values ($1, $2, current_timestamp, $3) returning *;",
+  };
+
   private getUserAuthQuery = {
-    text: "select public_address, name, nonce from " + userTableName + " where public_address = $1",
+    text:
+      "select public_address, name, nonce from " +
+      userTableName +
+      " where public_address = $1",
   };
 }
